# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server to create messages in postgres and kafka",
        "title": "Messagio Test Task API",
        "contact": {},
        "version": "1.0"
    },
    "host": "194.247.187.44:5000",
    "basePath": "/",
    "paths": {
        "/messages/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "Создает новое сообщение",
                "parameters": [
                    {
                        "description": "Сообщение",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/message.messageReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        },
        "/messages/stats": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "Выводит информацию о количестве необработанных сообщений",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "message.messageReq": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}